requirements.txt
tensorflow-cpu==2.16.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
plotly==5.22.0
streamlit==1.38.0
